requirements.txt
mypy==1.15.0
mypy_extensions==1.1.0
platformdirs==4.3.8
python-slugify==8.0.4
pytubefix==8.13.1
text-unidecode==1.3
typing_extensions==4.13.2
